requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.7.0
django-filter==24.3
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
drf-access-policy==1.5.0
drf-yasg==1.21.9
gis==0.2.1
inflection==0.5.1
packaging==24.2
pillow==11.0.0
psycopg==3.2.3
PyJWT==2.9.0
pyparsing==3.2.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
